==> notebooks/1-jmk-detailed fighter characteristics.ipynb <==
## Libraries and Globals

%matplotlib inline

import requests
import pandas as pd
import string
import time
from bs4 import BeautifulSoup
from pandas.plotting import scatter_matrix
import os 
import string
import random

FIGHTER_CHAR_BASE = "http://ufcstats.com/statistics/fighters?char="
FIGTHER_CHAR_END = "&page=all"
ALPHABET = list(string.ascii_lowercase)

row_ind = 1
all_fighters_char_urls = []
for letter in ALPHABET:
    url = FIGHTER_CHAR_BASE + letter + FIGTHER_CHAR_END
    time.sleep(random.uniform(0,1))
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    table = soup.find("table")
    rows = table.findAll("a")
    
    for row in rows:
        all_fighters_char_urls.append(row['href'])

all_fighters_char_urls = set(all_fighters_char_urls)

len(all_fighters_char_urls)

names = []
nick_names = []
fight_records = []
heights = []
weights = []
reaches = []
stances = []
DOBs = []

error_urls = []

url_num = 0

for url in all_fighters_char_urls:
    try:
        time.sleep(random.uniform(0,1))
        page = requests.get(url)
        soup = BeautifulSoup(page.content, 'html.parser')
    
        # Name
        names.append(soup.find('span',{'class':'b-content__title-highlight'}).text.replace('\n',"").strip())
        # Nickname
        nick_names.append(soup.find('p',{"class":"b-content__Nickname"}).text.replace('\n',"").strip())
        # Fight Record
        fight_records.append(soup.find('span',{"class","b-content__title-record"}).text.replace('\n',"").replace("Record: ","").strip())

        box = soup.find("div",{"class","b-list__info-box b-list__info-box_style_small-width js-guide"})
        stat_box = box.findAll("li")

        # Height
        heights.append(stat_box[0].text.split(":")[1].replace("\n","").strip())
        # Weight
        weights.append(stat_box[1].text.split(":")[1].replace("\n","").strip())
        # Reach
        reaches.append(stat_box[2].text.split(":")[1].replace("\n","").strip())
        # Stance
        stances.append(stat_box[3].text.split(":")[1].replace("\n","").strip())
        # DOB
        DOBs.append(stat_box[4].text.split(":")[1].replace("\n","").strip())
        
    except:
        error_urls.append(url)
        
    url_num += 1
    

fight_char_dict = {"name":names,
                  "nick_name":nick_names,
                  "fight_record":fight_records,
                  "height":heights,
                  "weight":weights,
                  "reach":reaches,
                  "stance":stances,
                  "DOB": DOBs}

for thing in fight_char_dict:
    print(len(fight_char_dict[thing]))

fight_char_df = pd.DataFrame(fight_char_dict)
fight_char_df

# save as csv
fight_char_df.to_csv("../data/fighter_char_v3.csv")

## Test Environment

test_url = "http://ufcstats.com/fighter-details/93fe7332d16c6ad9"
test_page = requests.get(test_url)
test_soup = BeautifulSoup(test_page.content, 'html.parser')

# nickname
test_soup.find('p',{"class":"b-content__Nickname"}).text.replace('\n',"").strip()

test_soup.find('span',{"class","b-content__title-record"}).text.replace('\n',"").replace("Record: ","").strip()

box = test_soup.find("div",{"class","b-list__info-box b-list__info-box_style_small-width js-guide"})
stat_box = box.findAll("li")

# Height
print(stat_box[0].text.split(":")[1].replace("\n","").strip())
# Weight
print(stat_box[1].text.split(":")[1].replace("\n","").strip())
# Reach
print(stat_box[2].text.split(":")[1].replace("\n","").strip())
# Stance
print(stat_box[3].text.split(":")[1].replace("\n","").strip())
# DOB
print(stat_box[4].text.split(":")[1].replace("\n","").strip())

for row in stat_box:
    print(row.text.replace('\n',''))
