==> notebooks/3-jmk-codebook.ipynb <==
import pandas as pd

stats_df = pd.read_csv("full_stats.csv")

variable_ids = list(stats_df.columns)

variable_labels = []
var_to_label = {
    "fighter": "fighter",
    "1": "1",
    "2": "2",
    "l": "landed",
    "a": "attempts",
    "f1": "fighter 1",
    "f2": "fighter 2",
    "r1": "round 1",
    "r2": "round 2",
    "r3": "round 3",
    "r4": "round 4",
    "r5": "round 5",
    "kds": "knockdowns",
    "sigstrikes": "significant strikes",
    "tds": "takedowns",
    "subs": "submissions",
    "revs": "reversals",
    "ss": "significant strikes",
    "head": "to the head",
    "body": "to the body",
    "leg": "to the leg",
    "distance": "from distance",
    "clinch": "from clinch",
    "ground": "from ground"
}

for col in variable_ids:
    col_vars = col.split("_")
    col_labels = []
    for var in col_vars:
        var_to_append = var_to_label[var] if var in var_to_label else var
        col_labels.append(var_to_append)
    label = " ".join(col_labels)
    variable_labels.append(label)

print(len(variable_labels))

    
#     if col == "fighter_1":
#         variable_labels.append("fighter 1")
#     elif col == "fighter_2":
#         variable_labels.append("fighter 2")
#     elif col == "winner":
#         variable_labels.append("winner")
#     elif(col[0:2] == "f1" or col[0:2] == "f2"):
#         temp_word = col.replace("l","landed").replace("a","attempts")
#         temp_word = temp_word.replace("f1","fighter 1").replace("f2","fighter 2")
#         temp_word = temp_word.replace("r1","round 1").replace("r2","round 2").replace("r3","round 3").replace("r4","round 4").replace("r5","round 5")
#         temp_word = temp_word.replace("kds", "knockdowns").replace("sigstrikes","significant strikes")
#         temp_word = temp_word.replace("tds", "takedowns").replace("subs","submissions").replace("rev","reversals")
#         temp_word = temp_word.replace("ss_head","significant strikes to the head").replace("ss_body","significant strikes to the body")
#         temp_word = temp_word.replace("ss_leg","significant strikes to the leg").replace("ss_distance","significant strikes from distance")
#         temp_word = temp_word.replace("ss_clinch","significant strikes from clinch").replace("ss_ground","significant strikes from ground").replace("_"," ")


#         variable_labels.append(temp_word)
    
#     else:
#         variable_labels.append(col)




code_book = pd.DataFrame({"Variable IDs":variable_ids, "Variable Labels": variable_labels})

code_book.to_csv("codebook.csv")
