==> notebooks/4.1-jmk-data wrangling data as is.ipynb <==
# 1.1 Import libraries, data, and globals

%matplotlib inline
%load_ext autoreload
%autoreload 2

# Get rid of warnings
import warnings
warnings.filterwarnings('ignore')

# General libraries
import requests
import pandas as pd
import string
import time
from bs4 import BeautifulSoup
from pandas.plotting import scatter_matrix
import os 
import string
import random
import numpy as np
import re
import datetime
from dateutil.relativedelta import relativedelta
import category_encoders as ce
import pickle

# Sklearn Specific
from sklearn.model_selection import cross_val_score 
from sklearn.model_selection import cross_validate
from sklearn import preprocessing

# sklearn algos
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn import svm
from sklearn.neighbors import KNeighborsClassifier 
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB # Naive Bayes
from sklearn.linear_model import Perceptron
from sklearn.linear_model import SGDClassifier # SGD
from sklearn.neural_network import MLPClassifier # neural network (multilayer perceptron)

# Bayesian Hyperparameter optimization
from hyperopt import hp, tpe, fmin

# Import data and remove Tony Johnson
fs_df = pd.read_csv("../data/scraped_data/full_stats_v5.csv")
fs_df = fs_df[fs_df['fighter_1'] != "Tony Johnson"]
fs_df = fs_df[fs_df['fighter_2'] != "Tony Johnson"]

fc_df = pd.read_csv("../data/scraped_data/fighter_char_v3.csv")
fc_df = fc_df[fc_df['name'] != "Tony Johnson"]

# 1.2 Functions

# Create dictionary with each fighter having a pd.df with all of their fights
def generate_fighters_dict():
    cols = two_pf.columns.tolist()
    fighters = {}
    for index, row in two_pf.iterrows():
        fighter = row["a_fighter"]
        if fighter not in fighters:
            fighters[fighter] = [row]
        else:
            fighters[fighter].append(row)
    for fighter in fighters:
        rows_list = fighters[fighter]
        fighters[fighter] = pd.DataFrame(columns=cols, data=rows_list)
    return fighters

# Gets filtered fighter information
def get_fighter_info(past_df, is_fighter, lb_num, is_exact):
    if is_exact:
        latest_df = past_df.head(lb_num)
    else:
        latest_df = past_df
        
    latest_lb_cum = latest_df.mean(skipna=True, numeric_only=True).to_dict()
    new_dict = {}
    if is_fighter:
        for var in latest_lb_cum:
            if var[0] == "f":
                new_dict[var] = latest_lb_cum[var]
    else:
        for var in latest_lb_cum:
            if var[0] == "f":
                new_var = var.replace("f_", "o_")
                new_dict[new_var] = latest_lb_cum[var]
    return(new_dict)

# Generates cumulative df
def generate_cum_df(fighters_dict, lb_num, is_exact):
    
    cumulative_df = pd.DataFrame(columns = two_pf.columns)
    
    for index, row in two_pf.iterrows():
        fighter_name = row["a_fighter"]
        opponent_name = row["a_opponent"]
        fight_date = row["date"]
        fighter_df = fighters_dict[fighter_name]
        opponent_df = fighters_dict[opponent_name]
        
        past_fighter_df = fighter_df.loc[fighter_df['date'] < fight_date]
        if past_fighter_df.shape[0] < lb_num:
            continue
        past_opponent_df = opponent_df.loc[opponent_df['date'] < fight_date]
        if past_opponent_df.shape[0] < lb_num:
            continue
        fighter_info = get_fighter_info(past_fighter_df, True, lb_num, is_exact)
        opponent_info = get_fighter_info(past_opponent_df, False, lb_num, is_exact)
        all_info = dict(fighter_info, **opponent_info)
        all_info['a_fighter'] = row['a_fighter']
        all_info['a_opponent'] = row['a_opponent']
        all_info['a_winner'] = row['a_winner']
        all_info['date'] = row['date']
        all_info['result'] = row['result']
    
        # Merge fight characterestics
        fc_index = 0
        f_init = "fo"
        fighters_list = [fighter_name,opponent_name]
        
        for fighter in fighters_list:
            fc_index_r = fc_index%2
            
            # Merge in height, reach, and stance
            all_info['{0}_height'.format(f_init[fc_index_r])] = fc_df.loc[fc_df['full_name'] == fighters_list[fc_index_r]]['height'].item()
            all_info['{0}_reach'.format(f_init[fc_index_r])] = fc_df.loc[fc_df['full_name'] == fighters_list[fc_index_r]]['reach'].item()
            all_info['{0}_stance'.format(f_init[fc_index_r])] = fc_df.loc[fc_df['full_name'] == fighters_list[fc_index_r]]['stance'].item()
            
            #Merge in age of fighter
            DOB = fc_df.loc[fc_df['full_name'] == fighters_list[fc_index_r]]['DOB'].item()
            if(DOB == "--"):
                break
            age = relativedelta(fight_date, DOB).years
            all_info['{0}_age'.format(f_init[fc_index_r])] = age

            fc_index += 1
            
        
        # Convert dict to pdf
        all_info_df = pd.DataFrame(all_info, index=[0])
        
        cumulative_df = pd.concat([cumulative_df, all_info_df])
        
    return(cumulative_df)


# 2.1 Data wrangling for full_stats

#### Add Totals

fs_df = fs_df.fillna(0) # Fill NaN with 0

# Calc time of fight in seconds
fs_df['duration'] = fs_df.apply(lambda x: (int(x['end_round'])-1)*60*5 + int(x['end_time'].split(":")[0])*60 + int(x['end_time'].split(":")[1]) , axis = 1)

# Convert date to date object
fs_df['date'] = fs_df.apply(lambda x: datetime.datetime.strptime(x['date'],'%Y-%m-%d').date(), axis = 1)


# Calc totals for each fighter
which_fighter = ["f1","f2"]
stat_name_ending = ["kds","sigstrikes_l","sigstrikes_a","strikes_l","strikes_a","tds_l","tds_a","subs_a","passes",
                   "revs","ss_head_l","ss_head_a","ss_body_l","ss_body_a","ss_leg_l","ss_leg_a","ss_dist_l","ss_dist_a",
                   "ss_clinch_l","ss_clinch_a","ss_ground_l","ss_ground_a"]

for f in which_fighter:
    
    for stat in stat_name_ending:
        
        fs_df[f+'_t_' + stat] = np.sum(fs_df.loc[:,(fs_df.columns.str.startswith(f) & fs_df.columns.str.endswith(stat))], axis =1)
        
        fs_df[f+'_pm_' + stat] =  fs_df[f+'_t_' + stat] / fs_df['duration'] * 60


#### Subset the Data for Key Stats and Per Minute 

matchers = ['fighter','winner',"_pm_","date"]
ml_cols = [s for s in fs_df.columns if any(xs in s for xs in matchers)]
ml_df = fs_df[ml_cols]

#### Make each fight have two rows, so that the label is win or lose

# Part 1
part_1_df = ml_df.copy()

part1_key_cols = part_1_df.columns
part1_value_cols = []

for col in part1_key_cols:
    new_col = col.replace("f1","f")
    new_col = new_col.replace("f2","o")
    new_col = new_col.replace("fighter_1","a_fighter")
    new_col = new_col.replace("fighter_2","a_opponent")
    new_col = new_col.replace("winner","a_winner")
    part1_value_cols.append(new_col)
    
p1_col_dict = dict(zip(part1_key_cols, part1_value_cols))

part_1_df = part_1_df.rename(columns = p1_col_dict)

# Part 2
part_2_df = ml_df.copy()

part2_key_cols = part_2_df.columns
part2_value_cols = []

for col in part2_key_cols:
    new_col = col.replace("f1","o")
    new_col = new_col.replace("f2","f")
    new_col = new_col.replace("fighter_1","a_opponent")
    new_col = new_col.replace("fighter_2","a_fighter")
    new_col = new_col.replace("winner","a_winner")
    part2_value_cols.append(new_col)
    
p2_col_dict = dict(zip(part2_key_cols, part2_value_cols))

part_2_df = part_2_df.rename(columns = p2_col_dict)

two_pf = pd.concat([part_1_df, part_2_df], sort = True)
two_pf = two_pf.sort_values(by ='date', ascending= False)
two_pf['result'] = two_pf.apply(lambda x: int(x['a_fighter'] == x['a_winner']), axis =1)


#### Show head and describe the wrangled dataset

two_pf.head(3)

two_pf.describe()

# 2.2 Data Wrangling for fighter_char

#### Calculate mean height and reach to fill in missing data

# Height
height_df = fc_df[fc_df['height'] != "--"]
height_df = height_df.reset_index()

for index, row in height_df.iterrows():
    height_list = row['height'].split(" ")
    p1 = int(re.sub("[^0-9]", "", height_list[0]))*12
    p2 = int(re.sub("[^0-9]", "", height_list[1]))
    height_df.loc[index, 'height'] = p1 + p2
    
mean_height = np.mean(height_df['height'])

# Reach
reach_df = fc_df[fc_df['reach'] != "--"]
reach_df = reach_df.reset_index()

for index, row in reach_df.iterrows():
    reach_df.loc[index, 'reach'] = int(row['reach'].replace('"',""))
    
mean_reach = np.mean(reach_df['reach'])

fc_df = fc_df.drop(['Unnamed: 0'], axis = 1) # drops first column

fc_df = fc_df.fillna(value = "") # fills na

# Change values to numerics and add full_name column
for index, row in fc_df.iterrows(): 
    # add full_name
    if(row['nick_name'] == ""):
        fc_df.loc[index,'full_name'] = row['name']
        
    else:
        fc_df.loc[index,'full_name'] = row['name'] + ' "' + row['nick_name'] + '"' 
    
    # convert height to int(inches)
    if(row['height'] != "--"):
        height_list = row['height'].split(" ")
        fc_df.loc[index,'height'] = int(re.sub("[^0-9]", "", height_list[0]))*12 + int(re.sub("[^0-9]", "", height_list[1]))
    else:
        fc_df.loc[index,'height'] = mean_height
    
    # convert reach to date object
    if(row['reach'] != "--"):
        fc_df.loc[index,'reach'] = int(row['reach'].replace('"',""))
    else:
        fc_df.loc[index,'reach'] = mean_reach
        
    # convert dob to date object
    if(row['DOB'] != "--"):
        date_str = row['DOB'].replace(",","").replace(" ","-")
        fc_df.loc[index,'DOB'] = datetime.datetime.strptime(date_str, '%b-%d-%Y').date()
        
        

#### Show head and describe the wrangled dataset

fc_df.head(3)

fc_df.describe()

# 3. Create cumulative statistics data with specified look back window

#### Create functions to make the look back window dataframe

fighters_dict = generate_fighters_dict() # Generates a dictionary with each fighter having his own pd.DataFrame()

## Make cumulative data using 1,3,5,7,10 lookback window

exact_1_df = generate_cum_df(fighters_dict,1, True)
exact_4_df = generate_cum_df(fighters_dict,4, True)  
exact_7_df = generate_cum_df(fighters_dict,7, True)  
exact_10_df = generate_cum_df(fighters_dict,10, True)

cumu_1_df = generate_cum_df(fighters_dict,1, False)    
cumu_4_df = generate_cum_df(fighters_dict,4, False)      
cumu_7_df = generate_cum_df(fighters_dict,7, False)
cumu_10_df = generate_cum_df(fighters_dict,10, False)

cumu_dfs_dict = {"Exact Data: 1 Fight Lookback Window" : exact_1_df,
                 "Exact Data: 4 Fight Lookback Window" : exact_4_df,
                 "Exact Data: 7 Fight Lookback Window" : exact_7_df,
                 "Exact Data: 10 Fight Lookback Window" : exact_10_df,
                 "Cumulative Data: 1 Fight Lookback Window": cumu_1_df,
                 "Cumulative Data: 4 Fight Lookback Window": cumu_4_df,
                 "Cumulative Data: 7 Fight Lookback Window": cumu_7_df,
                 "Cumulative Data: 10 Fight Lookback Window": cumu_10_df} #  create dictionary of the dfs

# 4. Feature Engineering (Merging in height, reach, stance)

fe_cumu_dfs_dict = cumu_dfs_dict.copy() # create copy of aggregate data sets

accuracy_stats = ['sigstrikes', 'strikes', 'tds', 'ss_head', 'ss_body', 'ss_leg', 'ss_dist', 'ss_clinch', 'ss_ground']
for key in fe_cumu_dfs_dict:
    
    df = fe_cumu_dfs_dict[key].copy()
    
    
    for stat in accuracy_stats:
        
        fe_index = 0
        f_init = "fo"
        
        for letter in f_init:
            
            fe_index_r = fe_index%2
            # Merge in height, reach, and stance
            df['{0}_acu_{1}'.format(f_init[fe_index_r], stat)] =  df['{0}_pm_{1}_l'.format(f_init[fe_index_r], stat)] / df['{0}_pm_{1}_a'.format(f_init[fe_index_r], stat)]

            fe_index += 1
            
    df = df.fillna(0)
    fe_cumu_dfs_dict[key] = df
    

# 5. Calculate Differentials

# 6. Save data for Machine Learning 

data_dicts_list = [cumu_dfs_dict,fe_cumu_dfs_dict] # add dictionaries to list

#### Save the dictionaries

with open('../data/temp_data/data_dicts_v3.pickle', 'wb') as handle:
    pickle.dump(data_dicts_list, handle, protocol=pickle.HIGHEST_PROTOCOL)

fe_cumu_lb1 = data_dicts_list[1]['Cumulative Data: 1 Fight Lookback Window']

fe_cumu_lb1.head(3)

fe_cumu_lb1.columns

# 7. Feature Selection? Use Johnson for reference
